# contracts/openapi/tenant-v1.yaml
openapi: 3.1.0
info:
  title: VPSFlow Tenant API
  version: 1.0.0
  description: Organization and project management for VPSFlow.
servers:
  - url: http://localhost:8080/api/v1
paths:
  /organizations:
    get:
      operationId: listOrganizations
      summary: List organizations for current user
      tags: [Tenant]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Organization list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizationList"

    post:
      operationId: createOrganization
      summary: Create a new organization
      tags: [Tenant]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrganizationRequest"
      responses:
        "201":
          description: Organization created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Organization:
      type: object
      required: [org_id, name, slug, created_at]
      properties:
        org_id:
          type: string
        name:
          type: string
        slug:
          type: string
        created_at:
          type: string
          format: date-time
    OrganizationList:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Organization"
    CreateOrganizationRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        slug:
          type: string
...